Add support for PreHash ML-DSA
This commit adds two functions crypto_sign_signature_pre_hash_internal
and crypto_sign_verify_pre_hash_internal implementing the pre-hashing mode of
ML-DSA. Instead of receiving a message, they receive a pre-hashed message.
Details can be found in Algorithm 4 and 5 in FIPS204.

The message to signed is formatted as
0x01 || ctxlen (1 byte) || ctx || oid || ph
where ph is the pre-hash and oid is the object identifier of the used hash
algorithm.

The ACVP client is adjusted to support the pre-hashing test cases. Note that
the ACVP testvectors only contain the message, not the pre-hash. I opted for
computing the hash in the ACVP Python client as for that we do not have to
add implementations for the 12 hash functions.

CBMC proofs for the new functions are added.

Resolves https://github.com/pq-code-package/mldsa-native/issues/39

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>