Build: Add EXTRA_MAKEFILE and EXTRA_SOURCES arguments
When EXTRA_MAKEFILE is passed, it will be included at the start of the main
Makefile. This can be used to configure builds requiring a large number of
flags (e.g., baremetal builds).

EXTRA_SOURCES can be passed to compile and link additional sources, e.g.,
platform files for baremetal builds.

Co-authored-by: Brendan Moran <brendan.moran@arm.com>
Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>