Allow additional CFLAGS for extra sources
Since the addition of -Wconversion and -Wsign-conversion to the default
CFLAGS, compilation of the CMSIS files for AN547 fails.

This commit allows custom Makefiles to not only provide EXTRA_SOURCES
to specify additional sources, but also EXTRA_SOURCES_CFLAGS for additional
CFLAGS, which in particular may overwrite default CFLAGS (they apply later).

This is leveraged in the case of m55-an547/platform.mk to disable
conversion and sign-conversion warnings for all CMSIS files.

Port of
https://github.com/pq-code-package/mlkem-native/pull/1220/commits/15fa58e9cf08da52038c68c283865bf750c1a7b1

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>