ACVP: Adjust to actual ACVP flow
This commit ports
https://github.com/pq-code-package/mlkem-native/pull/1052

With HashML-DSA support added in
https://github.com/pq-code-package/mldsa-native/pull/498, we now have full
coverage of the ACVP tests allowing us to switch to the actual ACVP flow:
In real ACVP validation, the internalProjection is not available.
Rather, one gets a prompt containing the inputs and has to produce a result
json that has the match the expected results hold by the ACVP server.

This commit modifies the acvp_client.py to follow this flow and aligns it with
the client in mlkem-native.
This in theory allows to perform CAVP validation without any changes, but I
have not done that yet.

Resolves
https://github.com/pq-code-package/mldsa-native/issues/294

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>