SLOTHY: Superoptimize ntt.S for the Neoverse N1
This commit adds the optimized backend dev/aarch64_opt. For now this backend
only differs from the clean backend in the NTT which is superoptimized using
SLOTHY for the Neoverse N1. For all other files it's a simple copy of the clean
backend. A Makefile is added that performs the optimization.
CI is adjusted to test both the clean and the opt backend.

The first loop of the NTT can be optimized in one go. The second loop is
too largeand we, hence, use the split heuristic.

I have experimented with the Cortex-A55 model as well - that results in
significantly faster code on A55, but results in a noticable slow down,
especially for A72 (see performance results in the pull request).
A72 performance seems more important than A55 performance.

I have experimented with applying some other optimizations (from the SLOTHY
paper):
 - Using st4 instead of the manual tranposition
 - Using scalar loads instead of vector loads
While those result in much better performance on Cortex-A55, they slow down
code on other platforms (see the pull request for details).

The autogen script is extended to allow running the optimization through the
--slothy flag.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>