Make value barrier volatile
This commit modifies the definition of the standard value barrier
by adding a `volatile` qualifier. Otherwise, there is risk the
compiler could correlate input and output merely from the fact that
the value barrier is deterministic. We have not seen this happening,
but it appears safer nontheless to add a volatile qualifier.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>