decompose: Remove separate input argument
To reduce stack usage, polyveck_decompose should be able to operate on one
of the outputs aliasing the input. This is already fully supported but
wasn't used to ease CBMC proofs/contracts at a cost of requiring additional
stack space which is undesirable.

This commit is an alternative to
https://github.com/pq-code-package/mldsa-native/pull/792
which extended the CBMC contracts to allow aliasing.

This commit instead, changes the signature of the decompose functions to
take 2 arguments instead of 3, where the second output also acts as and input.
The corresponding poly_ and native functions are adjusted accordingly.
CBMC contracts/proofs and unit tests are adjusted to the new signature.
The single callsite of polyveck_decompose is adjusted.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>